+++ nix/packages/default.nix
@@ -3,13 +3,19 @@
   advisory-db,
   craneLib,
   prefetchLanguagesFile,
+  prefetchLanguagesNickelFile,
 }:
 
 let
   binPkgs = callPackageNoOverrides ./bin.nix { };
 
   topiaryPkgs = callPackageNoOverrides ./topiary.nix {
-    inherit advisory-db craneLib prefetchLanguagesFile;
+    inherit
+      advisory-db
+      craneLib
+      prefetchLanguagesFile
+      prefetchLanguagesNickelFile
+      ;
   };
 in