+++ nix/utils/nickelUtils.nix
@@ -7,6 +7,10 @@
 
 let
   inherit (builtins)
+    concatStringsSep
+    isList
+    isString
+    isAttrs
     readFile
     toJSON
     fromJSON
@@ -14,6 +18,9 @@ let
     ;
   inherit (lib.strings)
     removeSuffix
+    ;
+  inherit (lib.attrsets)
+    mapAttrsToList
     ;
 
   /**
@@ -46,10 +53,34 @@ let
     in
     fromJSON (readFile "${jsonDrv}");
 
+  /**
+    Converts a Nix value to Nickel source syntax.
+
+    Strings are JSON-escaped, lists use `[...]`, and attribute sets use
+    `{ key = value, ... }` (Nickel record syntax).
+
+    # Type
+
+    ```
+    toNickelValue : Any -> String
+    ```
+  */
+  toNickelValue =
+    val:
+    if isList val then
+      "[${concatStringsSep ", " (map toNickelValue val)}]"
+    else if isString val then
+      toJSON val
+    else if isAttrs val then
+      "{ ${concatStringsSep ", " (mapAttrsToList (k: v: "${k} = ${toNickelValue v}") val)} }"
+    else
+      toString val;
+
 in
 {
   inherit
     fromNickelFile
     toJSONFile
+    toNickelValue
     ;
 }