+++ CHANGELOG.md
@@ -52,10 +52,10 @@ This name should be decided amongst the team before the release.
 ### Fixed
 - [#1176](https://github.com/topiary/topiary/pull/1176) Increase the stack size to 4MiB in Windows builds.
 
-<!--
 ### Added
-- <New feature>
+- [#1200](https://github.com/topiary/topiary/pull/1200) Build and deploy Topiary Docker images to ghcr.io.
 
+<!--
 ### Changed
 - <Changes in existing functionality>